backend/requirements.txt
pandas~=1.4.2
matplotlib
google-cloud-storage
flask
requests~=2.32.5